rename and simplify components
