+++ src/components/pages/App.tsx
@@ -1,10 +1,10 @@
 import "aos/dist/aos.css";
 import styled from "styled-components";
-import About from "../organisms/About";
-import Banner from "../organisms/Banner";
-import Contact from "../organisms/Contact";
-import Navbar from "../organisms/Navbar";
-import Portfolio from "../organisms/Portfolio";
+import About from "../sections/About";
+import Banner from "../sections/Banner";
+import Contact from "../sections/Contact";
+import Navbar from "../sections/Navbar";
+import Portfolio from "../sections/Portfolio";
 
 const StyledApp = styled.div`
   background: linear-gradient(to bottom, white, #ffb900);
