+++ src/components/sections/Banner/index.tsx
